requirements.txt
gradio==4.44.1
gradio_client
PyGithub
openai
requests
pandas
tiktoken
msal
rich
chevron
pypandoc
python-docx
markdown
matplotlib
pydantic
textdistance
scipy
transformers
lxml_html_clean
beautifulsoup4
playwright
fastapi
uvicorn
huggingface-hub==0.25.2
